+++ Verizon/src/main/java/com/mopub/mobileads/VerizonAdapterConfiguration.java
@@ -23,9 +23,9 @@ public class VerizonAdapterConfiguration extends BaseAdapterConfiguration {
 
     private static final String ADAPTER_VERSION = BuildConfig.VERSION_NAME;
     private static final String MOPUB_NETWORK_NAME = BuildConfig.NETWORK_NAME;
-    private static final String VAS_SITE_ID_KEY = "siteId";
 
-    static final String MEDIATOR_ID = "MoPubVAS-" + ADAPTER_VERSION;
+    public static final String MEDIATOR_ID = "MoPubVAS-" + ADAPTER_VERSION;
+    public static final String VAS_SITE_ID_KEY = "siteId";
 
     @NonNull
     @Override