+++ Verizon/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+  * 1.1.4.0
+    * Add support for rewarded video and native ad.
+    * This version of the adapters has been certified with Verizon 1.1.4.
+
   * 1.1.3.0
     * This version of the adapters has been certified with Verizon 1.1.3.